api/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.3
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.8.2
google-cloud-aiplatform==1.66.0
google-auth==2.29.0
requests==2.32.3
google-genai==0.5.0
Pillow==10.2.0
